+++ slo/src/Internal/SloTableContext.cs
@@ -22,7 +22,7 @@ public abstract class SloTableContext<T> : ISloContext
 {
     private const int IntervalMs = 100;
 
-    protected static readonly ILogger Logger = ISloContext.Factory.CreateLogger<SloTableContext<T>>();
+    private static readonly ILogger Logger = ISloContext.Factory.CreateLogger<SloTableContext<T>>();
 
     private volatile int _maxId;
 
@@ -129,8 +129,7 @@ public abstract class SloTableContext<T> : ISloContext
         Logger.LogInformation("Run task is finished");
         return;
 
-        async Task ShootingTask(RateLimiter rateLimitPolicy, string operationType,
-            Func<T, RunConfig, Task<int>> action)
+        async Task ShootingTask(RateLimiter rateLimitPolicy, string operationType, Func<T, RunConfig, Task> action)
         {
             var metricFactory = Metrics.WithLabels(new Dictionary<string, string>
                 {
@@ -182,11 +181,6 @@ public abstract class SloTableContext<T> : ISloContext
                 }
             );
 
-            var retryAttempts = metricFactory.CreateGauge(
-                "sdk_retry_attempts",
-                "Current retry attempts, categorized by operation type."
-            );
-
             var pendingOperations = metricFactory.CreateGauge(
                 "sdk_pending_operations",
                 "Current number of pending operations, categorized by type."
@@ -218,9 +212,8 @@ public abstract class SloTableContext<T> : ISloContext
                         var sw = Stopwatch.StartNew();
                         try
                         {
-                            var attempts = await action(client, runConfig);
+                            await action(client, runConfig);
                             sw.Stop();
-                            retryAttempts.Set(attempts);
                             operationsTotal.Inc();
                             pendingOperations.Dec();
                             operationsSuccessTotal.Inc();
@@ -248,7 +241,7 @@ public abstract class SloTableContext<T> : ISloContext
     // return attempt count & StatusCode operation
     protected abstract Task<int> Save(T client, SloTable sloTable, int writeTimeout);
 
-    protected abstract Task<(int, object?)> Select(T client, (Guid Guid, int Id) select, int readTimeout);
+    protected abstract Task<object?> Select(T client, (Guid Guid, int Id) select, int readTimeout);
 
     protected abstract Task<int> SelectCount(T client);
 
@@ -272,12 +265,10 @@ public abstract class SloTableContext<T> : ISloContext
         return Save(client, sloTable, config.WriteTimeout);
     }
 
-    private async Task<int> Select(T client, RunConfig config)
+    private async Task Select(T client, RunConfig config)
     {
         var id = Random.Shared.Next(_maxId);
-        var (attempts, _) = await Select(client, new ValueTuple<Guid, int>(GuidFromInt(id), id), config.ReadTimeout);
-
-        return attempts;
+        _ = await Select(client, new ValueTuple<Guid, int>(GuidFromInt(id), id), config.ReadTimeout);
     }
 
     private static Guid GuidFromInt(int value)