+++ src/Ydb.Sdk/src/Ado/Session/ImplicitSession.cs
@@ -52,6 +52,5 @@ internal class ImplicitSession : ISession
 
     public void Close() => _source.ReleaseLease();
 
-    private static YdbException NotSupportedTransaction =>
-        new(StatusCode.BadRequest, "Transactions are not supported in implicit sessions");
+    private static YdbException NotSupportedTransaction => new("Transactions are not supported in implicit sessions");
 }